feat: towards adding support for building container
 - added a parser that builds an AST for Gunfile
    - Gunfile is a new file format/DSL I made up, similar to Dockerfile
 - added a "build" command
 - running the commands in the Exec blocks in Gunfile works
 - but the image isn't saved because archiver dies when there's a
   nonexisting symlink

+++ main.go
@@ -1,10 +1,14 @@
 package main
 
 import (
-	"log"
+	"encoding/base64"
+	"fmt"
 	"io/ioutil"
+	"log"
 	"os"
 	"os/exec"
+	"strconv"
+	"strings"
 	"syscall"
 
 	"github.com/mholt/archiver/v3"
@@ -26,6 +30,10 @@ func main() {
 		run()
 	case "child":
 		child()
+	case "childMultiple":
+		childMultiple()
+	case "build":
+		build()
 	default:
 		log.Fatalln("Unknown command", os.Args[1])
 	}
@@ -50,17 +58,35 @@ func run() {
 		log.Fatalln("Gib command to run")
 	}
 
-	log.Println("Hmm, running", os.Args[3:])
+	Run(os.Args[2], os.Args[3], os.Args[4:])
+}
+
+func Run(imagePath string, command string, args []string) {
+	log.Printf("Hmm, running %v with args %v", command, args)
+
+	dname, err := RunNoClean(imagePath, command, args, "child")
+	if dname != "" {
+		defer os.RemoveAll(dname)
+	}
+
+	if err != nil {
+		log.Fatalln(err)
+	}
+}
+
+// runs image
+// returns the new root directory, which is not cleaned up
+// childCommand must be "child" or "childMultiple"
+func RunNoClean(imagePath string, command string, args []string, childCommand string) (string, error) {
 
 	// extract root fs and do stuff
-	dname, err := setupNewRoot(os.Args[2])
+	dname, err := setupNewRoot(imagePath)
 	if err != nil {
-		log.Fatalln("Error extracting root FS:", err)
+		return dname, fmt.Errorf("Error extracting root FS: %v", err)
 	}
 	// log.Println("Extracted root FS to", dname)
-	defer os.RemoveAll(dname)
 
-	cmd := exec.Command(os.Args[0], append([]string{"child", dname}, os.Args[3:]...)...)
+	cmd := exec.Command(os.Args[0], append([]string{childCommand, dname, command}, args...)...)
 
 	cmd.Stdin = os.Stdin
 	cmd.Stdout = os.Stdout
@@ -69,7 +95,7 @@ func run() {
 
 	// namespaces
 	cmd.SysProcAttr = &syscall.SysProcAttr{
-		Cloneflags: syscall.CLONE_NEWUTS | syscall.CLONE_NEWPID | syscall.CLONE_NEWNS | syscall.CLONE_NEWUSER,
+		Cloneflags:   syscall.CLONE_NEWUTS | syscall.CLONE_NEWPID | syscall.CLONE_NEWNS | syscall.CLONE_NEWUSER,
 		Unshareflags: syscall.CLONE_NEWNS,
 		Credential: &syscall.Credential{
 			Uid: 0,
@@ -77,21 +103,23 @@ func run() {
 		},
 		UidMappings: []syscall.SysProcIDMap{{
 			ContainerID: 0,
-			HostID: os.Geteuid(),
-			Size: 1,
+			HostID:      os.Geteuid(),
+			Size:        1,
 		}},
 		GidMappings: []syscall.SysProcIDMap{{
 			ContainerID: 0,
-			HostID: os.Getegid(),
-			Size: 1,
+			HostID:      os.Getegid(),
+			Size:        1,
 		}},
 	}
 
 	err = cmd.Run()
 
 	if err != nil {
-		log.Fatalln("Error running child command:", err)
+		return dname, fmt.Errorf("Error running child command: %v", err)
 	}
+
+	return dname, nil
 }
 
 func child() {
@@ -111,5 +139,60 @@ func child() {
 
 	if err != nil {
 		log.Fatalln("Error running command:", err)
+	}
+}
+
+// each string of cmdWithArgs must be the command and its arguments separated by \u0000
+func RunMultipleNoClean(imagePath string, cmdWithArgs []string) (string, error) {
+	log.Printf("Hmm, running commands: %v", cmdWithArgs)
+
+	encodedCmdWithArgs := make([]string, len(cmdWithArgs))
+	for i, cmd := range cmdWithArgs {
+		encoded := base64.StdEncoding.EncodeToString([]byte(cmd))
+		encodedCmdWithArgs[i] = encoded
+	}
+
+	dname, err := RunNoClean(imagePath, encodedCmdWithArgs[0], encodedCmdWithArgs[1:], "childMultiple")
+
+	return dname, err
+}
+
+func childMultiple() {
+	// TODO: somehow get multiple command-args here and run them individually
+	//       also do the chroot stuff before that
+
+	syscall.Sethostname([]byte("guntainer"))
+	syscall.Chroot(os.Args[2])
+	syscall.Chdir("/")
+	syscall.Mount("proc", "proc", "proc", 0, "")
+	defer syscall.Unmount("/proc", 0)
+
+	for _, encoded := range os.Args[3:] {
+		decoded, err := base64.StdEncoding.DecodeString(encoded)
+		cmdWithArgsString := string(decoded)
+
+		divider, err := strconv.Unquote("'\u0000'")
+		if err != nil {
+			log.Fatalf("Divider could not be unquoted. If you're seeing this error, uhhh. You shoudn't be.")
+		}
+		cmdWithArgs := strings.Split(cmdWithArgsString, divider)
+
+		if err != nil {
+			log.Fatalf("Could not decode command %v: %v", encoded, err)
+		}
+
+		log.Printf("Running command %v with args: %v", cmdWithArgs[0], cmdWithArgs[1:])
+
+		cmd := exec.Command(cmdWithArgs[0], cmdWithArgs[1:]...)
+
+		cmd.Stdin = os.Stdin
+		cmd.Stdout = os.Stdout
+		cmd.Stderr = os.Stderr
+
+		err = cmd.Run()
+
+		if err != nil {
+			log.Fatalf("Error running command %v: %v", cmdWithArgs, err)
+		}
 	}
 }
